chore(setup): enhance plugin setup with goreleaser and deployment fixes
- Update Makefile to deploy plugin with binary, plugin.yaml, and install-binary.sh
- Refine goreleaser configuration for correct YAML structure and build setup
- Ensure install-binary.sh script handles plugin installation effectively
