--- .github/.goreleaser.yml
@@ -1,63 +1,0 @@
-project_name: kanvas-snapshot
-version: 2
-before:
-  hooks:
-    - go mod tidy
-
-builds:
-  - main: cmd/kanvas-snapshot/main.go
-
-    env:
-      - CGO_ENABLED=0
-
-    ldflags:
-      - -s -w
-      - -X main.providerToken={{.Env.PROVIDER_TOKEN}}  
-      - -X main.MesheryCloudApiBaseUrl="https://meshery.layer5.io"
-      - -X main.MesheryApiBaseUrl="https://playground.meshery.io"
-
-    goos:
-      - darwin
-      - linux
-      - windows
-
-    goarch:
-      - 386
-      - amd64
-      - arm
-      - arm64
-
-    ignore:
-      - goos: windows
-        goarch: arm
-      - goos: windows
-        goarch: arm64
-
-archives:
-  - id: stable
-    name_template: >-
-      {{ .ProjectName }}_{{.Version}}_
-      {{- title .Os }}_
-      {{- if eq .Arch "amd64" }}x86_64
-      {{- else if eq .Arch "386" }}i386
-      {{- else }}{{ .Arch }}{{ end }}
-    format: tar.gz
-    format_overrides:
-      - goos: windows
-        format: zip
-
-checksum:
-  name_template: 'checksums.txt'
-
-snapshot:
-  version_template: "{{ .Tag }}-next"
-
-changelog:
-  sort: asc
-  filters:
-    exclude:
-      - '^docs:'
-      - '^test:'
-
-release:
-  name_template: "Helm Kanvas Snapshot {{.Tag}}"
