+++ .goreleaser.yml
@@ -1,0 +1,53 @@
+project_name: kanvas-snapshot
+version: 2
+before:
+  hooks:
+    - go mod tidy
+
+builds:
+  - env:
+      - CGO_ENABLED=0
+    ldflags:
+      - -s -w
+      - -X main.providerToken={{.Env.PROVIDER_TOKEN}}
+      - -X main.MesheryCloudApiBaseUrl="https://meshery.layer5.io"
+      - -X main.MesheryApiBaseUrl="https://playground.meshery.io"
+    goos:
+      - linux
+      - darwin
+      - freebsd
+      - windows
+    goarch:
+      - amd64
+      - '386'
+      - arm
+      - arm64
+    binary: '{{ .ProjectName }}'
+
+archives:
+  - format: tar.gz
+    name_template: >-
+      {{- .ProjectName }}_
+      {{- .Version }}_
+      {{- .Os }}_
+      {{- if eq .Arch "amd64" }}x86_64
+      {{- else if eq .Arch "386" }}i386
+      {{- else }}{{ .Arch }}{{ end }}
+    files:
+      - plugin.yaml
+      - LICENSE
+      - completion.yaml
+      - install-binary.sh
+
+checksum:
+  name_template: 'checksums.txt'
+
+snapshot:
+  version_template: "{{ .Tag }}-next"
+
+changelog:
+  sort: asc
+  filters:
+    exclude:
+      - '^docs:'
+      - '^test:'
