+++ artifacthub-repo.yml
@@ -1,0 +1,19 @@
+# Artifact Hub repository metadata file
+#
+# Some settings like the verified publisher flag or the ignored packages won't
+# be applied until the next time the repository is processed. Please keep in
+# mind that the repository won't be processed if it has not changed since the
+# last time it was processed. Depending on the repository kind, this is checked
+# in a different way. For Helm http based repositories, we consider it has
+# changed if the `index.yaml` file changes. For git based repositories, it does
+# when the hash of the last commit in the branch you set up changes. This does
+# NOT apply to ownership claim operations, which are processed immediately.
+#
+repositoryID: 1b4e2e51-4979-43c4-a164-c26426c196b2
+owners: # (optional, used to claim repository ownership)
+  - name: meshery
+    email: maintainers@meshery.io
+#ignore: # (optional, packages that should not be indexed by Artifact Hub)
+#  - name: package1
+#  - name: package2 # Exact match
+#    version: beta # Regular expression (when omitted, all versions are ignored)
