+++ main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	cmd "github.com/meshery/helm-kanvas-snapshot/cmd/kanvas-snapshot"
+)
+
+func main() {
+	cmd.Main()
+}
